requirements.txt
fastapi
uvicorn[standard]
python-telegram-bot==20.6
python-dotenv
tinydb
